pyproject.toml
[project]
name = "voice-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit>=1.1.2",
    "livekit-agents[turn-detector]>=1.4.6",
    "livekit-plugins-deepgram>=1.4.6",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.4.6",
    "livekit-plugins-silero>=1.4.6",
    "python-dotenv>=1.2.2",
]
